Refactor Resume component into component file Resume
