+++ src/components/app/Resume/index.js
@@ -1,0 +1,3 @@
+import Resume from './Resume';
+
+export default Resume;
